Add last tests inside the parameters dialog (before transform it to SWT); Use String constants in VariantsUtils; Fix *.properties copy bug and the "&&" mistake.

+++ org.but4reuse.adapters.eclipse/src/org/but4reuse/adapters/eclipse/generator/utils/FileAndDirectoryUtils.java
@@ -99,7 +99,7 @@ public class FileAndDirectoryUtils {
 	 * @throws IOException if the copy is impossible
 	 */
 	public static void deleteFile(File file) throws IOException {
-		if(file != null & file.exists()){
+		if(file != null && file.exists()){
 			FileUtils.forceDelete(file);
 			return;
 		}
